Add tests to WallTest and PlayerBoardTest to check for color flip bug, #128
Added unit test to WallTest that ensures that placeTileInRowAndColor() works correctly.
Added unit test to PlayerBoardTest that places a non-red tile in the top row of the stairs,
then processes end of round, and asserts the correct color and column in the correct row
are marked as occupied by the wall.